Updated Sign-In, Sign-Up, Admin - User Switch
