+++ src/app/components/NavBar.tsx
@@ -1,40 +1,61 @@
 import React from 'react';
- 
+import { useAuth } from '../contexts/AuthContext';
+import Link from 'next/link';
+
 interface NavBarProps {
   links: { name: string, path: string }[];
-  setCurrentPage: (page: string) => void;
 }
- 
-const NavBar: React.FC<NavBarProps> = ({ links, setCurrentPage }) => {
-  // Filter out the sign-up link
-  const filteredLinks = links.filter(link => link.name !== 'Sign Up');
- 
+
+const NavBar: React.FC<NavBarProps> = ({ links }) => {
+  const { role, setRole } = useAuth();
+
+  const handleRoleSwitch = () => {
+    setRole(role === 'admin' ? 'user' : 'admin');
+  };
+
+  const getFilteredLinks = () => {
+    const commonLinks = links.filter(link => link.name !== 'Sign Up');
+    if (role === 'admin') {
+      return commonLinks.map(link => ({
+        ...link,
+        path: link.path === '/' ? '/' : `/admin${link.path}`
+      }));
+    }
+    return commonLinks;
+  };
+
+  const filteredLinks = getFilteredLinks();
+
   return (
     <>
       <nav className="flex justify-between items-center p-4 bg-white shadow-md">
         <div className="flex items-center">
-          <a href="/" onClick={() => setCurrentPage('/')}>
-            <img src="logo.png" alt="Logo" className="h-12 w-55" /> 
-          </a>
+          <Link href="/">
+            <img src="/logo.png" alt="Logo" className="h-12 w-55 cursor-pointer" />
+          </Link>
         </div>
-        <div className="flex items-center space-x-8 ml-auto"> 
-          <ul className="flex space-x-8"> 
+        <div className="flex items-center space-x-8 ml-auto">
+          <ul className="flex space-x-8">
             {filteredLinks.map(link => (
               <li key={link.path}>
-                <button
-                  className="text-lg text-black hover:text-blue-600 focus:text-blue-600 hover:font-bold focus:font-bold transition-transform duration-300"
-                  onClick={() => setCurrentPage(link.path)}
-                >
-                  {link.name}
-                </button>
+                <Link href={link.path}>
+                  <span className="text-lg text-black hover:text-blue-600 focus:text-blue-600 hover:font-bold focus:font-bold transition-transform duration-300">
+                    {link.name}
+                  </span>
+                </Link>
               </li>
             ))}
           </ul>
         </div>
+        <div className="ml-4">
+          <button onClick={handleRoleSwitch} className="text-lg text-black bg-gray-200 p-2 rounded">
+            {role === 'admin' ? 'Switch to User' : 'Switch to Admin'}
+          </button>
+        </div>
       </nav>
-      <div className="h-1 bg-gray-900 shadow-md"></div> 
+      <div className="h-1 bg-gray-900 shadow-md"></div>
     </>
   );
 };
- 
+
 export default NavBar;
