+++ src/app/layout.tsx
@@ -1,22 +1,32 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+"use client";
 import "./globals.css";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+import { ThemeProvider } from "./contexts/ThemeContext";
+import { AuthProvider } from "./contexts/AuthContext";
+import NavBar from './components/NavBar';
+import Footer from './components/Footer';
 
 export default function RootLayout({
   children,
-}: Readonly<{
+}: {
   children: React.ReactNode;
-}>) {
+}) {
+  const links = [
+    { name: 'Home', path: '/' },
+    { name: 'Jobs', path: '/Jobs' },
+    { name: 'Profile', path: '/Profile' },
+    { name: 'Sign In', path: '/SignIn' },
+  ];
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <ThemeProvider>
+        <AuthProvider>
+          <body>
+            <NavBar links={links} />
+            {children}
+            <Footer />
+          </body>
+        </AuthProvider>
+      </ThemeProvider>
     </html>
   );
 }
